+++ persistent/Database/Persist/Sql/Orphan/PersistQuery.hs
@@ -99,7 +99,7 @@ instance PersistQueryRead SqlBackend where
                 [] -> ""
                 ords -> " ORDER BY " <> T.intercalate "," ords
         cols = commaSeparated . entityColumnNames t
-        sql conn = connLimitOffset conn (limit,offset) (not (null orders)) $ mconcat
+        sql conn = connLimitOffset conn (limit,offset) $ mconcat
             [ "SELECT "
             , cols conn
             , " FROM "
@@ -120,7 +120,7 @@ instance PersistQueryRead SqlBackend where
         wher conn = if null filts
                     then ""
                     else filterClause False conn filts
-        sql conn = connLimitOffset conn (limit,offset) (not (null orders)) $ mconcat
+        sql conn = connLimitOffset conn (limit,offset) $ mconcat
             [ "SELECT "
             , cols conn
             , " FROM "
@@ -419,8 +419,12 @@ orderClause includeTable conn o =
         $ connEscapeFieldName conn (fieldName x)
 
 -- | Generates sql for limit and offset for postgres, sqlite and mysql.
-decorateSQLWithLimitOffset::Text -> (Int,Int) -> Bool -> Text -> Text
-decorateSQLWithLimitOffset nolimit (limit,offset) _ sql =
+decorateSQLWithLimitOffset
+    :: Text
+    -> (Int,Int)
+    -> Text
+    -> Text
+decorateSQLWithLimitOffset nolimit (limit,offset) sql =
     let
         lim = case (limit, offset) of
                 (0, 0) -> ""