+++ persistent/Database/Persist/SqlBackend/Internal/MkSqlBackend.hs
@@ -74,7 +74,7 @@ data MkSqlBackendArgs = MkSqlBackendArgs
     -- ^ A tag displaying what database the 'SqlBackend' is for. Can be
     -- used to differentiate features in downstream libraries for different
     -- database backends.
-    , connLimitOffset :: (Int,Int) -> Bool -> Text -> Text
+    , connLimitOffset :: (Int,Int) -> Text -> Text
     -- ^ Attach a 'LIMIT/OFFSET' clause to a SQL query. Note that
     -- LIMIT/OFFSET is problematic for performance, and indexed range
     -- queries are the superior way to offer pagination.